% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{aggregate_to_water_weeks}
\alias{aggregate_to_water_weeks}
\title{aggregate_to_water_weeks}
\usage{
aggregate_to_water_weeks(x)
}
\arguments{
\item{x}{tibble to be aggregated}
}
\value{
data in water weeks
}
\description{
aggregates daily data to water weeks
}
