% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{back_calc_missing_flows}
\alias{back_calc_missing_flows}
\title{back_calc_missing_flows}
\usage{
back_calc_missing_flows(x, compute_from = "i")
}
\arguments{
\item{x}{tibble of s, i, r in water weeks}

\item{compute_from}{variable (i, r) from which to back calculate}
}
\value{
data ready for horizon search
}
\description{
back_calc_missing_flows
}
\details{
assumes no losses (evap., seepage, etc.)
}
