% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizon_search_daily.R
\name{compute_fcast_short_scores}
\alias{compute_fcast_short_scores}
\title{compute_fcast_short_scores}
\usage{
compute_fcast_short_scores(
  dam,
  max_fill_gap = 3,
  data_dir = NULL,
  cutoff_year = NULL,
  min_allowable_points = 10,
  method = "spearman"
)
}
\arguments{
\item{dam}{name of dam to be analyzed}

\item{max_fill_gap}{maximum gap (days) allowed for storage and inflow data}

\item{min_allowable_points}{minimum number of points for building r-a function (default = 10)}

\item{water_week}{water week}

\item{horizon}{look ahead horizon (1 <= h <= 52)}

\item{compute_from}{variable (i, r) from which to back calculate}
}
\value{
tibble of release versus availability for given water week and horizon
}
\description{
compute_fcast_short_scores
}
\details{
compute availability for given week of year and horizon (in weeks)
}
