% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{convert_to_complete_water_years}
\alias{convert_to_complete_water_years}
\title{convert_to_complete_water_years}
\usage{
convert_to_complete_water_years(x)
}
\arguments{
\item{x}{tibble to be filtered}
}
\value{
filtered data in water years
}
\description{
convert_to_complete_water_years
}
