% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{.log_env}
\alias{.log_env}
\title{HORIZON - Logging Functions}
\usage{
.log_env
}
\description{
Functions for managing console output logging to files.
Duplicates stdout to both screen and log file (does not redirect).
Modelled on GAIA's ELEUTHIA logging.
}
