% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{fill_NAs}
\alias{fill_NAs}
\title{fill_NAs}
\usage{
fill_NAs(x, max_fill_gap = 10)
}
\arguments{
\item{x}{tibble to be gap-filled}

\item{max_fill_gap}{maximum gap (days) allowed for storage and inflow data}
}
\value{
daily, gap-filled reservoir data in metric volumes (Mm^3)
}
\description{
fills NA values using a smoothing spline
}
