% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizon_search.R
\name{get_optimized_models}
\alias{get_optimized_models}
\title{get_optimized_models}
\usage{
get_optimized_models(dam, all_valid_combos, write_to = NULL,
  water_week = NULL, horizon = NULL, compute_from = "i",
  max_fill_gap = 10, write_loc = NULL, data_dir = NULL,
  cutoff_year = NULL, opt_mode = "two_param")
}
\arguments{
\item{dam}{name of dam to be analyzed}
}
\value{
optimized piecewise functions for all water week and horizon combinations
}
\description{
get_optimized_models
}
\details{
gets all model parameters (and associated r-sq values)
}
