% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizon_search_daily.R
\name{get_optimized_models_daily}
\alias{get_optimized_models_daily}
\title{get_optimized_models_daily}
\usage{
get_optimized_models_daily(
  dam,
  all_valid_combos,
  write_to = NULL,
  mth = NULL,
  horizon = NULL,
  max_fill_gap = 10,
  max_horizon = 7,
  write_loc = NULL,
  data_dir = NULL,
  cutoff_year = NULL,
  opt_mode = "two_param"
)
}
\arguments{
\item{dam}{name of dam to be analyzed}
}
\value{
optimized piecewise functions for all water week and horizon combinations
}
\description{
get_optimized_models_daily
}
\details{
gets all model parameters (and associated r-sq values)
}
