% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{hplot_calculated_flows}
\alias{hplot_calculated_flows}
\title{hplot_calculated_flows}
\usage{
hplot_calculated_flows(dam, plot_years = NULL)
}
\arguments{
\item{dam}{name of dam to be analyzed}
}
\value{
plot of i and r observed versus computed
}
\description{
hplot_calculated_flows
}
