% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{hplot_fcast_short}
\alias{hplot_fcast_short}
\title{hplot_fcast_short}
\usage{
hplot_fcast_short(
  dam,
  mth,
  horizon,
  max_fill_gap = 3,
  data_dir = NULL,
  cutoff_year = NULL,
  min_allowable_points = 5
)
}
\arguments{
\item{dam}{dam}

\item{mth}{month}

\item{horizon}{horizon}

\item{max_fill_gap}{maximum number of NA days to fill across}

\item{data_dir}{data directory}

\item{cutoff_year}{cutoff year}

\item{min_allowable_points}{minimum allowable points for model build}
}
\value{
plot of selected horizon for all water weeks
}
\description{
hplot_fcast_short
}
\details{
plot short forecast model
}
