% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{hplot_fcast_short_gof}
\alias{hplot_fcast_short_gof}
\title{hplot_fcast_short_gof}
\usage{
hplot_fcast_short_gof(
  dam,
  max_fill_gap = 3,
  data_dir = NULL,
  cutoff_year = NULL,
  min_allowable_points = 5,
  method = "spearman",
  min_score = 0.7
)
}
\arguments{
\item{dam}{dam}

\item{max_fill_gap}{maximum number of NA days to fill across}

\item{data_dir}{data directory}

\item{cutoff_year}{cutoff year}

\item{min_allowable_points}{minimum allowable points for model build}
}
\value{
plot of selected horizon for all water weeks
}
\description{
hplot_fcast_short_gof
}
\details{
plot r2 resulrs for all models
}
