% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{hplot_model_performances}
\alias{hplot_model_performances}
\title{hplot_model_performances}
\usage{
hplot_model_performances(file_loc, smoothing = "x_x_x_x")
}
\arguments{
\item{file_loc}{file containing piecewise functions to be plotted}

\item{smoothing}{vector of values in order r_sq cutoff, smoothing_span, "", ""}
}
\value{
plots of horizon and associated r_sq for all weeks
}
\description{
hplot_model_performances
}
