% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{hplot_ready_data}
\alias{hplot_ready_data}
\title{hplot_ready_data}
\usage{
hplot_ready_data(dam, water_week, horizon, add_piecewise_fn = F,
  compute_from = "i", max_fill_gap = 10)
}
\arguments{
\item{dam}{name of dam to be analyzed}
}
\value{
plot of r as function of availability
}
\description{
hplot_ready_data
}
