% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{hplot_selected_models}
\alias{hplot_selected_models}
\title{hplot_model_performances}
\usage{
hplot_selected_models(optimized_models, smooth = FALSE)
}
\arguments{
\item{optimized_models}{tibble containing all piecewise functions for all water weeks}

\item{smooth}{logical. Is smoothing desired? (default: FALSE)}
}
\value{
plot of selected horizon for all water weeks
}
\description{
hplot_model_performances
}
