% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{read_dam}
\alias{read_dam}
\title{read_dam}
\usage{
read_dam(dam, data_dir = NULL)
}
\arguments{
\item{dam}{name of dam to be loaded into the environment}
}
\value{
tibble of observed dam data (storage, inflow, release)
}
\description{
reads in observed dam data as tibble
}
