% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{read_dam_USRDATS}
\alias{read_dam_USRDATS}
\title{read_dam_USRDATS}
\usage{
read_dam_USRDATS(grand_id, data_dir)
}
\arguments{
\item{dam}{GRanD ID of dam}
}
\value{
tibble of observed dam data (storage, inflow, release)
}
\description{
reads in observed dam data as tibble
}
