% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{select_all_models}
\alias{select_all_models}
\title{select_all_models}
\usage{
select_all_models(results_dir, write_loc = NULL)
}
\arguments{
\item{results_dir}{directory of results (optimized models for all weeks)}

\item{write_loc}{location to write results}
}
\value{
list of selected models
}
\description{
select_all_models
}
