% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{show_dams}
\alias{show_dams}
\title{show_dams}
\usage{
show_dams(data_source, data_dir = NULL)
}
\arguments{
\item{data_source}{one of "usbr", "usbrpn", "cdec", "usgs", "usace", "twdb"}
}
\value{
list of dam names
}
\description{
shows all dams from a given source
}
