% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizon_search.R
\name{compute_availability}
\alias{compute_availability}
\title{compute_availability}
\usage{
compute_availability(dam, water_week, horizon, max_fill_gap = 10,
  compute_from = "i")
}
\arguments{
\item{dam}{name of dam to be analyzed}

\item{water_week}{water week}

\item{horizon}{look ahead horizon (1 <= h <= 52)}

\item{max_fill_gap}{maximum gap (days) allowed for storage and inflow data}

\item{compute_from}{variable (i, r) from which to back calculate}
}
\value{
tibble of release versus availability for given water week and horizon
}
\description{
compute_availability
}
\details{
compute availability for given week of year and horizon (in weeks)
}
