% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{convert_to_metric}
\alias{convert_to_metric}
\title{convert_to_metric}
\usage{
convert_to_metric(x)
}
\arguments{
\item{x}{tibble to be converted}
}
\value{
daily reservoir data in metric volumes (Mm^3)
}
\description{
convert_to_metric
}
\details{
converts daily data (s in af; i, r in cfs) metric volumes (Mm^3)
}
