% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\docType{data}
\name{.log_env}
\alias{.log_env}
\title{HORIZON - Logging Functions}
\format{
An object of class \code{environment} of length 4.
}
\usage{
.log_env
}
\description{
Functions for managing console output logging to files.
Duplicates stdout to both screen and log file (does not redirect).
Modelled on GAIA's ELEUTHIA logging.
}
\keyword{datasets}
