\name{horizon-package}
\alias{horizon-package}
\alias{horizon}
\docType{package}
\title{Horizon Search Algorithm}
\description{Implementation of the horizon search algorithm and sky view factor algorithms of Dozier and Frew (1990) for a raster digital terrain model.}
\details{
\tabular{ll}{
Package: \tab horizon\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-09-09\cr
Depends: \tab raster\cr
License: \tab GPL-2\cr
}
}
\author{
Jasper Van doninck

Maintainer: Jasper Van doninck <vandoninck.jasper@gmail.com>
}
\references{
Dozier, J. and Frew, J. (1990) Rapid calculation of terrain parameters for radiation modeling from digital elevation data, \emph{IEEE Transactions on Geoscience and Remote Sensing} 5, 963--969.
}