% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizon-package.R
\docType{package}
\name{horizon-package}
\alias{horizon}
\alias{horizon-package}
\title{horizon: Inflow forecast horizons used in reservoir release decisions.}
\description{
Estimates inflow forecast horizons used in reservoir release decisions.
}
\author{
\strong{Maintainer}: Sean Turner \email{sean.turner@pnnl.gov}

}
\keyword{internal}
