% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizon.R
\name{horizon}
\alias{horizon}
\title{Horizon chart}
\usage{
horizon(dates, df, width = 800, height = 600, axis_height = 30,
  axis_ticks = 4)
}
\arguments{
\item{dates}{Vector of dates as character strings \code{"YYYY-MM-DD"}.}

\item{df}{Data frame with rows = dates, columns = values to plot}

\item{width}{width in pixels}

\item{height}{height in pixels}

\item{axis_height}{height of axis in pixels}

\item{axis_ticks}{Number of ticks on axes}
}
\description{
Horizon chart of multiple time series
}
\examples{
file <- system.file("extdata", "test.csv", package="horizon")
x <- read.csv(file, stringsAsFactors=FALSE)
horizon(x[,1], x[,-1])

}

