% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{hplot_selected_models}
\alias{hplot_selected_models}
\title{hplot_selected_models}
\usage{
hplot_selected_models(selected_models)
}
\arguments{
\item{optimized_models}{tibble containing all piecewise functions for all water weeks}

\item{smooth}{logical. Is pre-smoothing desired? (default: FALSE)}
}
\value{
plot of selected horizon for all water weeks
}
\description{
hplot_selected_models
}
