% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{select_best_horizon}
\alias{select_best_horizon}
\title{select_best_horizon}
\usage{
select_best_horizon(set, pre_smooth = FALSE, post_smooth = TRUE)
}
\arguments{
\item{set}{input set}

\item{smooth}{logical. Is smoothing desired? (default: FALSE)}
}
\value{
smoothed horizons
}
\description{
identify horizon that produces best-fit model
}
